+++ main-service/src/main/java/ru/practicum/main/service/request/model/ConfirmedRequests.java
@@ -1,4 +1,4 @@
 package ru.practicum.main.service.request.model;
 
-public record ConfirmedRequests(long eventId, long countRequests) {
+public record ConfirmedRequests(long eventId, int countRequests) {
 }
